pyproject.toml
[tool.poetry]
name = "api-client-template"
version = "0.1.0"
description = ""
authors = ["Aristide <aristide.bamazi@imsp-uac.org>"]
license = "mit"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.264"
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
